pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshai-sdk"
version = "0.1.0"
description = "SDK for MeshAI - AI Agent Interoperability Platform"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MeshAI Labs", email = "dev@meshai.dev"},
]
keywords = ["ai", "agents", "interoperability", "langchain", "crewai", "autogen"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "python-jose[cryptography]>=3.3.0",
    "websockets>=11.0.0",
    "prometheus-client>=0.17.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.10",
]
crewai = ["crewai>=0.1.0"]
autogen = ["pyautogen>=0.2.0"]
anthropic = ["anthropic>=0.21.0"]
google = [
    "google-generativeai>=0.4.0",
    "google-cloud-aiplatform>=1.42.0",
]
amazon = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
]
openai = ["openai>=1.12.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
all = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.10",
    "crewai>=0.1.0",
    "pyautogen>=0.2.0",
    "anthropic>=0.21.0",
    "google-generativeai>=0.4.0",
    "google-cloud-aiplatform>=1.42.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "openai>=1.12.0",
]

[project.scripts]
meshai = "meshai.cli:main"

[project.urls]
Homepage = "https://meshai.dev"
Documentation = "https://docs.meshai.dev"
Repository = "https://github.com/meshailabs/meshai-sdk"
"Bug Tracker" = "https://github.com/meshailabs/meshai-sdk/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true